+++ tests/CopyFile002.hs
@@ -9,9 +9,9 @@ main _t = do
   -- Similar to CopyFile001 but moves a file in the current directory
   -- (Bug #1652 on GHC Trac)
   writeFile from contents
-  T(expectEq) () [from] . sort =<< getDirectoryContentsA "."
+  T(expectEq) () [from] . sort =<< listDirectory "."
   copyFile from to
-  T(expectEq) () [from, to] . sort =<< getDirectoryContentsA "."
+  T(expectEq) () [from, to] . sort =<< listDirectory "."
   T(expectEq) () contents =<< readFile to
   where
     contents = "This is the data\n"